src/main/java/cruise/umple/sample/downloader/consistent/Consistents.java
/**
 * 
 */
package cruise.umple.sample.downloader.consistent;

import java.io.ByteArrayOutputStream;
import java.io.IOException;

import com.fasterxml.jackson.databind.ObjectMapper;

/**
 * @author Kevin Brightwell <kevin.brightwell2@gmail.com>
 *
 * @since 11 Mar 2015
 */
public abstract class Consistents {

  private Consistents() { }
  
  public static String toJson(ImportRepositorySet repository) {
    ObjectMapper mapper = new ObjectMapper();
    try (ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
      mapper.writeValue(baos, repository);
      
      return baos.toString();
    } catch (IOException e) {
      throw new IllegalStateException(e);
    }    
  }
}
